+++ internal/provider/data_source_iosxr_router_bgp_vrf_test.go
@@ -22,6 +22,23 @@ func TestAccDataSourceIosxrRouterBGPVRF(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "timers_bgp_holdtime", "20"),
 					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "bfd_minimum_interval", "10"),
 					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "bfd_multiplier", "4"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.neighbor_address", "10.1.1.2"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.remote_as", "65002"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.description", "My Neighbor Description"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.ignore_connected_check", "true"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.ebgp_multihop_maximum_hop_count", "10"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.bfd_minimum_interval", "10"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.bfd_multiplier", "4"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.local_as", "65003"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.local_as_no_prepend", "true"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.local_as_replace_as", "true"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.local_as_dual_as", "true"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.password", "12341C2713181F13253920"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.shutdown", "false"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.timers_keepalive_interval", "5"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.timers_holdtime", "20"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.update_source", "GigabitEthernet0/0/0/1"),
+					resource.TestCheckResourceAttr("data.iosxr_router_bgp_vrf.test", "neighbors.0.ttl_security", "false"),
 				),
 			},
 		},
@@ -39,6 +56,25 @@ resource "iosxr_router_bgp_vrf" "test" {
 	timers_bgp_holdtime = "20"
 	bfd_minimum_interval = 10
 	bfd_multiplier = 4
+	neighbors = [{
+		neighbor_address = "10.1.1.2"
+		remote_as = "65002"
+		description = "My Neighbor Description"
+		ignore_connected_check = true
+		ebgp_multihop_maximum_hop_count = 10
+		bfd_minimum_interval = 10
+		bfd_multiplier = 4
+		local_as = "65003"
+		local_as_no_prepend = true
+		local_as_replace_as = true
+		local_as_dual_as = true
+		password = "12341C2713181F13253920"
+		shutdown = false
+		timers_keepalive_interval = 5
+		timers_holdtime = "20"
+		update_source = "GigabitEthernet0/0/0/1"
+		ttl_security = false
+	}]
 }
 
 data "iosxr_router_bgp_vrf" "test" {